backend/requirements.txt
# Core API
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.27.0,<0.29.0
python-multipart>=0.0.6

# Config
pydantic>=2.6.0,<2.8.0
pydantic-settings>=2.2.0,<2.4.0

# Database
sqlalchemy>=2.0.25,<2.1.0
psycopg2-binary>=2.9.9
alembic>=1.13.1,<1.14.0

# PDF Processing
PyMuPDF>=1.23.26,<1.25.0
pdfplumber>=0.11.0,<0.12.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.2.0

# Vector DB
qdrant-client>=1.7.0,<1.9.0

# Embeddings
sentence-transformers>=2.3.1,<2.7.0


torch>=2.6.0,<2.8.0
safetensors>=0.4.2
# Graph
networkx>=3.2.1,<3.4.0

# Ollama + HTTP (FIXED CONFLICT HERE)
ollama>=0.1.7,<0.2.0
httpx>=0.25.2,<0.26.0
google-generativeai>=0.8.3,<0.9.0

# Utilities 
python-dotenv>=1.0.0
numpy>=1.26.0,<2.0.0